initial build of GVFS provider for Linux
This is largely a duplicate of the macOS build tooling and
platform-specific code, with some changes to ensure we build
successfully on Linux (e.g., testing for "!= Windows" is no longer
sufficient to determine that the platform is OS X).

Note that for now the BuildGVFSForLinux.sh script exits early,
without checking if a native libprojfs library is available,
performing any installation or packaging, etc.

The GVFS.sln configuration changes should ensure that the
ProjFS.Linux.Managed and GVFS.Platform.Linux subprojects
are built on all platforms (and the Mac and Windows equivalents
are built on Linux), so as to allow the GVFS.UnitTests project
to build on Linux, as it seems to do multi-platform unit testing.

However, we don't yet require the GVFS.Mount.Linux project
to build on other platforms, and the GVFS.FunctionalTests project's
configuration and depedencies will also need to be revisited
once the Linux code is actually functional.